# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Esta es una API generada automáticamente con Swaggo.",
        "title": "nbox API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Norlis Viamonte",
            "url": "http://www.example.com/support",
            "email": "norlis.viamonte@gmail.com"
        },
        "version": "2.0"
    },
    "basePath": "/",
    "paths": {
        "/api/auth/token": {
            "post": {
                "description": "authentication token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Token",
                "parameters": [
                    {
                        "description": "Payload",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_auth.TokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token generated successfully",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "token": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    }
                }
            }
        },
...
        "/api/prefix": {
            "get": {
                "description": "Get the list of configurations",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "prefix-config"
                ],
                "summary": "List Available configs",
                "responses": {
                    "200": {
                        "description": "List of prefix config",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_prefix.Config"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    }
                },
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/prefix/backends": {
            "get": {
                "description": "Get the list of supported storage backends",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "prefix-config"
                ],
                "summary": "List Available Backend Types",
                "responses": {
                    "200": {
                        "description": "List of backend types (e.g. dynamodb, parameterstore)",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    }
                },
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/prefix/resolve": {
            "get": {
                "description": "plain value",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "prefix-config"
                ],
                "summary": "Prefix config",
                "parameters": [
                    {
                        "type": "string",
                        "description": "prefix path",
                        "name": "v",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_prefix.Config"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    }
                },
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
        "/api/track/key": {
            "get": {
                "description": "history changes",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tracking"
                ],
                "summary": "History",
                "parameters": [
                    {
                        "type": "string",
                        "description": "key path",
                        "name": "v",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Start time. Supports: 30m, 2h, 1.5d, 2w, 1M or ISO-8601",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End time. Supports: 30m, 2h, 1.5d, 2w, 1M or ISO-8601",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_tracking.Record"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    }
                },
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ]
            }
        }
    },
    "definitions": {
        "internal_auth.TokenRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "internal_prefix.Config": {
            "type": "object",
            "properties": {
                "prefix": {
                    "description": "PK",
                    "type": "string"
                },
                "tags": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "typeAllowed": {
                    "description": "TODO: Deprecated",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_prefix.StorageBackendType"
                    }
                },
                "typeDefault": {
                    "$ref": "#/definitions/internal_prefix.StorageBackendType"
                },
                "typeSecure": {
                    "$ref": "#/definitions/internal_prefix.StorageBackendType"
                }
            }
        },
        "internal_prefix.StorageBackendType": {
            "type": "string",
            "enum": [
                "dynamodb",
                "parameterstore",
                "parameterstore_secure"
            ],
            "x-enum-varnames": [
                "BackendDynamoDB",
                "BackendParameterStore",
                "BackendParameterStoreSecure"
            ]
        },
        "internal_tracking.Record": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "hash": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "secure": {
                    "type": "boolean"
                },
                "storageBackend": {
                    "$ref": "#/definitions/nbox_internal_prefix.StorageBackendType"
                },
                "updatedAt": {
                    "type": "string"
                },
                "updatedBy": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
...
        "nbox_internal_prefix.StorageBackendType": {
            "type": "string",
            "enum": [
                "dynamodb",
                "parameterstore",
                "parameterstore_secure"
            ],
            "x-enum-varnames": [
                "BackendDynamoDB",
                "BackendParameterStore",
                "BackendParameterStoreSecure"
            ]
        },
        "problem.Detail": {
            "type": "object",
            "properties": {
                "detail": {
                    "description": "Detail is a human-readable explanation specific to this occurrence.",
                    "type": "string"
                },
                "instance": {
                    "description": "Instance is a URI identifying the specific occurrence of the problem.",
                    "type": "string"
                },
                "requestId": {
                    "description": "RequestID correlates the problem to a specific request.",
                    "type": "string"
                },
                "status": {
                    "description": "Status is the HTTP status code for this occurrence.",
                    "type": "integer"
                },
                "timestamp": {
                    "description": "Timestamp records when the problem was created.",
                    "type": "string"
                },
                "title": {
                    "description": "Title is a short, human-readable summary of the problem type.\nShould not change across occurrences of the same problem.",
                    "type": "string"
                },
                "type": {
                    "description": "Type is a URI identifying the problem type. Should provide\ndeveloper-readable documentation.",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "BearerAuth": {
            "description": "Bearer token authentication. Enter your JWT token in the format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}